Factor out http-get-query-util.tsx.
